Refactor analytic effects into page related effect files
