+++ src/pages/health-declaration/health-declaration.module.ts
@@ -2,6 +2,9 @@ import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 import { ComponentsModule } from '../../components/components.module';
 import { HealthDeclarationPage } from './health-declaration';
+import { EffectsModule } from '@ngrx/effects';
+import { HealthDeclarationAnalyticsEffects } from './health-declaration.analytics.effects';
+import { AnalyticsProvider } from '../../providers/analytics/analytics';
 
 @NgModule({
   declarations: [
@@ -9,7 +12,11 @@ import { HealthDeclarationPage } from './health-declaration';
   ],
   imports: [
     IonicPageModule.forChild(HealthDeclarationPage),
+    EffectsModule.forFeature([HealthDeclarationAnalyticsEffects]),
     ComponentsModule,
   ],
+  providers: [
+    AnalyticsProvider    
+  ]
 })
 export class HealthDeclarationPageModule {}
