+++ src/pages/pass-finalisation/pass-finalisation.module.ts
@@ -2,6 +2,9 @@ import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 import { ComponentsModule } from '../../components/components.module';
 import { PassFinalisationPage } from './pass-finalisation';
+import { EffectsModule } from '@ngrx/effects';
+import { PassFinalisationAnalyticsEffects } from './pass-finalisation.analytics.effects';
+import { AnalyticsProvider } from '../../providers/analytics/analytics';
 
 @NgModule({
   declarations: [
@@ -9,7 +12,11 @@ import { PassFinalisationPage } from './pass-finalisation';
   ],
   imports: [
     IonicPageModule.forChild(PassFinalisationPage),
+    EffectsModule.forFeature([PassFinalisationAnalyticsEffects]),
     ComponentsModule,
   ],
+  providers: [
+    AnalyticsProvider    
+  ]
 })
 export class PassFinalisationPageModule {}
