+++ src/pages/test-report/test-report.module.ts
@@ -2,6 +2,9 @@ import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 import { ComponentsModule } from '../../components/components.module';
 import { TestReportPage } from './test-report';
+import { EffectsModule } from '@ngrx/effects';
+import { TestReportAnalyticsEffects } from './test-report.analytics.effects';
+import { AnalyticsProvider } from '../../providers/analytics/analytics';
 
 @NgModule({
   declarations: [
@@ -9,7 +12,11 @@ import { TestReportPage } from './test-report';
   ],
   imports: [
     IonicPageModule.forChild(TestReportPage),
+    EffectsModule.forFeature([TestReportAnalyticsEffects]),
     ComponentsModule,
   ],
+  providers: [
+    AnalyticsProvider
+  ]
 })
 export class TestReportPageModule {}
