+++ src/pages/waiting-room/waiting-room.module.ts
@@ -2,6 +2,9 @@ import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 import { ComponentsModule } from '../../components/components.module';
 import { WaitingRoomPage } from './waiting-room';
+import { EffectsModule } from '@ngrx/effects';
+import { WaitingRoomAnalyticsEffects } from './waiting-room.analytics.effects';
+import { AnalyticsProvider } from '../../providers/analytics/analytics';
 
 @NgModule({
   declarations: [
@@ -9,7 +12,11 @@ import { WaitingRoomPage } from './waiting-room';
   ],
   imports: [
     IonicPageModule.forChild(WaitingRoomPage),
+    EffectsModule.forFeature([WaitingRoomAnalyticsEffects]),
     ComponentsModule,
   ],
+  providers: [
+    AnalyticsProvider
+  ]
 })
 export class WaitingRoomPageModule {}
